updated swiftlint errors and made linter strict
