+++ .github/workflows/swiftlint.yml
@@ -14,4 +14,6 @@ jobs:
     steps:
       - uses: actions/checkout@v1
       - name: GitHub Action for SwiftLint
-        uses: norio-nomura/action-swiftlint@3.1.0
+        uses: norio-nomura/action-swiftlint@3.2.1
+        with:
+          args: --strict
